requirements.txt
flask
Flask-WTF
Flask_SQLAlchemy
bootstrap-flask
markdown
pynacl
bleach
gunicorn
